Move format price utility to misc folder
