+++ src/components/product/ProductCard/index.jsx
@@ -1,6 +1,6 @@
 import { isDiscounted } from "../../../utils/product/isDiscounted";
 import { getDiscountPercent } from "../../../utils/product/getDiscountPercent";
-import { formatPrice } from "../../../utils/product/formatPrice";
+import { formatPrice } from "../../../utils/misc/formatPrice";
 import Button from "../../ui/Button";
 
 export default function ProductCard({ product }) {
